Refactor UserImageView environment initialization
