+++ GliaWidgets/Sources/View/Chat/GVA/Gallery/GalleryList/GvaGalleryListView.swift
@@ -23,12 +23,7 @@ final class GvaGalleryListView: BaseView {
 
     private lazy var operatorImageView = UserImageView(
         with: props.style.operatorImage,
-        environment: .init(
-            data: environment.data,
-            uuid: environment.uuid,
-            gcd: environment.gcd,
-            imageViewCache: environment.imageViewCache
-        )
+        environment: .create(with: environment)
     )
     lazy var collectionView = SelfSizingCollectionView(
         frame: .zero,
