+++ GliaWidgets/Sources/View/Chat/GVA/GvaResponseTextView.swift
@@ -7,12 +7,7 @@ final class GvaResponseTextView: ChatMessageView {
                 guard operatorImageView == nil else { return }
                 let operatorImageView = UserImageView(
                     with: viewStyle.operatorImage,
-                    environment: .init(
-                        data: environment.data,
-                        uuid: environment.uuid,
-                        gcd: environment.gcd,
-                        imageViewCache: environment.imageViewCache
-                    )
+                    environment: .create(with: environment)
                 )
                 self.operatorImageView = operatorImageView
                 operatorImageViewContainer.addSubview(operatorImageView)
